backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
geopandas
Fiona
pyproj
shapely
pandas
GeoAlchemy2